Changed pathing to :game/product

+++ db/migrate/20260408204726_drop_admins_table.rb
@@ -1,5 +1,5 @@
 class DropAdminsTable < ActiveRecord::Migration[8.1]
   def change
-    drop_table :admins
+    drop_table :admins if table_exists?(:admins)
   end
 end
